Fix linting and remove cyclical dependencies
