+++ packages/nodes-base/nodes/EmailReadImap/test/v2/EmailReadImapV2.node.test.ts
@@ -1,6 +1,8 @@
+// eslint-disable-next-line n8n-nodes-base/node-filename-against-convention
 import { mock } from 'jest-mock-extended';
-import { INodeTypeBaseDescription, ITriggerFunctions } from 'n8n-workflow';
-import { ICredentialsDataImap } from '../../../../credentials/Imap.credentials';
+import { type INodeTypeBaseDescription, type ITriggerFunctions } from 'n8n-workflow';
+
+import { type ICredentialsDataImap } from '../../../../credentials/Imap.credentials';
 import { EmailReadImapV2 } from '../../v2/EmailReadImapV2.node';
 
 jest.mock('@n8n/imap', () => {
